chore 📦: Update dependencies and add new package in pyproject.toml
This commit updates several dependencies and adds a new package 'typing-extensions'. The changes include updating the version of 'rich' from 13.7.0 to 14.0.0, adding 'typing-extensions' under the 'lint' group, and updating the 'test' group in 'tool.pdm.dev-dependencies' to include 'pytest-cov' and 'pylint'. Additionally, a new script 'docker-ctp' is added to the 'scripts' section of the project configuration. These changes enhance the project's dependencies, improve code quality tools, and add a new feature for running Docker images with enhanced functionality.